# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vezir Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/missions/{mission_id}": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get Mission",
        "operationId": "get_mission_api_v1_missions__mission_id__get",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetailResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/missions/{mission_id}/cancel": {
      "post": {
        "tags": [
          "mission-mutations"
        ],
        "summary": "Cancel Mission",
        "description": "Cancel a running/active mission (D-090: destructive — requires confirmation).\n\nValid states: pending, planning, executing, gate_check, rework, approval_wait.\nCompleted/failed/aborted/timed_out → 409.",
        "operationId": "cancel_mission_api_v1_missions__mission_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/retry": {
      "post": {
        "tags": [
          "mission-mutations"
        ],
        "summary": "Retry Mission",
        "description": "Retry a failed/aborted mission (OD-10: creates new mission via controller).\n\nValid states: failed, aborted, timed_out.\nRunning/pending/completed → 409.",
        "operationId": "retry_mission_api_v1_missions__mission_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/pause": {
      "post": {
        "tags": [
          "mission-mutations"
        ],
        "summary": "Pause Mission",
        "description": "Pause a running mission — pauses after current stage completes.\n\nValid states: pending, planning, executing, gate_check, rework, approval_wait.\nAlready paused / completed / failed → 409.",
        "operationId": "pause_mission_api_v1_missions__mission_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/resume": {
      "post": {
        "tags": [
          "mission-mutations"
        ],
        "summary": "Resume Mission",
        "description": "Resume a paused mission.\n\nValid states: paused.\nRunning / completed / failed → 409.",
        "operationId": "resume_mission_api_v1_missions__mission_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/skip-stage": {
      "post": {
        "tags": [
          "mission-mutations"
        ],
        "summary": "Skip Stage",
        "description": "Skip the current stage in a running mission.\n\nValid states: pending, planning, executing, gate_check, rework, approval_wait.\nPaused / completed / failed → 409.",
        "operationId": "skip_stage_api_v1_missions__mission_id__skip_stage_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard/missions/{mission_id}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Dashboard Mission",
        "description": "Get full mission detail.",
        "operationId": "get_dashboard_mission_api_v1_dashboard_missions__mission_id__get",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIError",
        "description": "Standard error response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionDetailResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "mission": {
            "$ref": "#/components/schemas/MissionSummary"
          }
        },
        "type": "object",
        "required": [
          "mission"
        ],
        "title": "MissionDetailResponse",
        "description": "Wrapper for single mission detail."
      },
...
      "MutationResponse": {
        "properties": {
          "requestId": {
            "type": "string",
            "title": "Requestid"
          },
          "lifecycleState": {
            "type": "string",
            "title": "Lifecyclestate",
            "default": "requested"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "requestedAt": {
            "type": "string",
            "title": "Requestedat"
          },
          "acceptedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptedat"
          },
          "appliedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliedat"
          },
          "rejectedReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejectedreason"
          },
          "timeoutAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeoutat"
          }
        },
        "type": "object",
        "required": [
          "requestId",
          "targetId",
          "requestedAt"
        ],
        "title": "MutationResponse",
        "description": "D-096 mutation response contract.\n\nAPI always returns lifecycleState=requested.\nSubsequent states (accepted/applied/rejected/timed_out) via SSE only."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}